+++ src/compounds/legacy-product-table/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@1.7.0...@uswitch/trustyle.legacy-product-table@1.7.1) (2021-08-26)
+
+
+### Bug Fixes
+
+* clickable row with secondary link ([fee0f65](https://github.com/uswitch/trustyle/commit/fee0f65))
+
+
+
+
+
+# [1.7.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@1.6.0...@uswitch/trustyle.legacy-product-table@1.7.0) (2021-08-25)
+
+
+### Features
+
+* add secondary cta to legacy pt ([220db84](https://github.com/uswitch/trustyle/commit/220db84))
+
+
+
+
+
 # [1.6.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@1.5.3...@uswitch/trustyle.legacy-product-table@1.6.0) (2021-08-18)